fix(runtime)!: remove cumulative execution limits across Flame
Upgrade Scope agent to v0.27.0 and align Run/Goal execution, durable
settlement, recovery, and all CLI/Desktop consumers with unbounded
long-running work. Retain resource concurrency and capacity controls.
Remove obsolete contracts, storage fields, UI states, tests, comments,
and documentation without compatibility adapters.

Validation: CLI workspace suite passed; Desktop 2437 tests passed and
10 targeted browser checks passed. Runtime contract generation and
focused lifecycle checks passed. Existing failures remain in Runtime
cancellation prefix preservation and two Desktop HTTP tool-persistence
cases (reproduced with pre-cleanup tests). Desktop aggregate checks also
report existing type, formatting, and unused-export failures.

BREAKING CHANGE: Run/Goal cumulative budget fields and budget terminal
states are removed from APIs and persistence. Consumers must use the
current contract. Standalone CLI dependency publication remains pending.

+++ runtime/internal/adapter/agentexec/interaction_invocation_identity.go
@@ -20,7 +20,7 @@ func modelInvocationID(invocation interaction.ModelInvocation) (runtimeidentity.
 	return modelInvocationIDFrom(invocation.EffectID(), invocation.ModelCallSequence())
 }
 
-func modelInvocationIDFrom(effectID agent.EffectID, modelCallSequence uint32) (runtimeidentity.EffectID, error) {
+func modelInvocationIDFrom(effectID agent.EffectID, modelCallSequence uint64) (runtimeidentity.EffectID, error) {
 	digest := sha256.New()
 	_, _ = digest.Write([]byte(effectID.String()))
 	_, _ = digest.Write([]byte{0})
@@ -39,7 +39,7 @@ func toolInvocationID(invocation interaction.ToolInvocation) (runtimeidentity.Ef
 
 func logicalToolCallID(
 	caller agent.ProcessID,
-	modelCallSequence uint32,
+	modelCallSequence uint64,
 	toolCallIndex uint32,
 	sourceID string,
 	name string,
@@ -52,10 +52,10 @@ func logicalToolCallID(
 	_, _ = digest.Write([]byte(sourceID))
 	_, _ = digest.Write([]byte{0})
 	_, _ = digest.Write([]byte(name))
-	return parsedInvocationID(toolInvocationNamespace, digest.Sum(nil), toolCallIndex)
+	return parsedInvocationID(toolInvocationNamespace, digest.Sum(nil), uint64(toolCallIndex))
 }
 
-func parsedInvocationID(namespace string, digest []byte, ordinal uint32) (runtimeidentity.EffectID, error) {
+func parsedInvocationID(namespace string, digest []byte, ordinal uint64) (runtimeidentity.EffectID, error) {
 	return runtimeidentity.ParseEffect(
 		namespace + ":" + hex.EncodeToString(digest) + ":" + strconv.FormatUint(uint64(ordinal), 10),
 	)
